macro_ai_agent/vision/processor.py
import cv2
import numpy as np
from pathlib import Path
from typing import Dict, List, Tuple
import easyocr
import logging

class VisionProcessor:
    def __init__(self):
        self.logger = logging.getLogger('VisionProcessor')
        self.reader = easyocr.Reader(['en'])
        
    def process_image(self, image_path: Path) -> Dict:
        """Process an image and extract visual requirements"""
        try:
            # Read image
            image = cv2.imread(str(image_path))
            if image is None:
                raise ValueError(f"Could not read image: {image_path}")

            # Extract features
            features = {
                "layout": self.analyze_layout(image),
                "colors": self.extract_colors(image),
                "text": self.extract_text(image),
                "components": self.detect_components(image)
            }

            return features

        except Exception as e:
            self.logger.error(f"Error processing image: {str(e)}")
            raise

    def analyze_layout(self, image: np.ndarray) -> Dict:
        """Analyze the layout structure of the image"""
        # Convert to grayscale
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        
        # Edge detection
        edges = cv2.Canny(gray, 50, 150)
        
        # Find contours
        contours, _ = cv2.findContours(edges, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
        
        # Analyze layout structure
        layout = {
            "sections": [],
            "grid_structure": self.detect_grid(contours),
            "symmetry": self.analyze_symmetry(contours)
        }
        
        return layout

    def extract_colors(self, image: np.ndarray) -> Dict:
        """Extract dominant colors and color scheme"""
        # Convert to RGB
        rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
        
        # Reshape image
        pixels = rgb.reshape(-1, 3)
        
        # Use K-means to find dominant colors
        from sklearn.cluster import KMeans
        kmeans = KMeans(n_clusters=5)
        kmeans.fit(pixels)
        
        # Get color palette
        colors = kmeans.cluster_centers_.astype(int)
        
        return {
            "dominant_colors": colors.tolist(),
            "color_scheme": self.classify_color_scheme(colors)
        }

    def extract_text(self, image: np.ndarray) -> List[Dict]:
        """Extract text from the image using OCR"""
        # Use EasyOCR to detect text
        results = self.reader.readtext(image)
        
        # Process results
        text_elements = []
        for (bbox, text, prob) in results:
            if prob > 0.5:  # Confidence threshold
                text_elements.append({
                    "text": text,
                    "confidence": prob,
                    "position": bbox
                })
        
        return text_elements

    def detect_components(self, image: np.ndarray) -> List[Dict]:
        """Detect UI components in the image"""
        # TODO: Implement component detection using computer vision
        # This could include:
        # - Buttons
        # - Input fields
        # - Navigation bars
        # - Cards
        # - Images
        return []

    def detect_grid(self, contours: List) -> Dict:
        """Detect grid structure in the layout"""
        # TODO: Implement grid detection
        return {
            "type": "unknown",
            "columns": 0,
            "rows": 0
        }

    def analyze_symmetry(self, contours: List) -> Dict:
        """Analyze symmetry in the layout"""
        # TODO: Implement symmetry analysis
        return {
            "horizontal_symmetry": 0.0,
            "vertical_symmetry": 0.0
        }

    def classify_color_scheme(self, colors: np.ndarray) -> str:
        """Classify the color scheme of the image"""
        # TODO: Implement color scheme classification
        return "unknown" 